+++ recipes/dd-agent.rb
@@ -17,6 +17,14 @@
 # limitations under the License.
 #
 
+# Fail here at converge time if no api_key is set
+ruby_block 'datadog-api-key-unset' do
+  block do
+    raise "Set ['datadog']['api_key'] as an attribute or on the node's run_state to configure this node's Datadog Agent."
+  end
+  only_if { Chef::Datadog.api_key(node).nil? }
+end
+
 is_windows = node['platform_family'] == 'windows'
 
 # Install the agent
@@ -42,7 +50,7 @@ directory node['datadog']['config_dir'] do
   else
     owner 'dd-agent'
     group 'root'
-    mode 0755
+    mode '755'
   end
 end
 
@@ -51,27 +59,25 @@ end
 # To add integration-specific configurations, add 'datadog::config_name' to
 # the node's run_list and set the relevant attributes
 #
-raise "Add a ['datadog']['api_key'] attribute to configure this node's Datadog Agent." if node['datadog'] && node['datadog']['api_key'].nil?
 
-api_keys = [node['datadog']['api_key']]
-dd_urls = [node['datadog']['url']]
-node['datadog']['extra_endpoints'].each do |_, endpoint|
-  next unless endpoint['enabled']
-  api_keys << endpoint['api_key']
-  dd_urls << if endpoint['url']
-               endpoint['url']
-             else
-               node['datadog']['url']
-             end
-end
-
-extra_config = {}
-node['datadog']['extra_config'].each do |option, value|
-  next if value.nil?
-  extra_config[option] = value
-end
-
-template agent_config_file do
+template agent_config_file do # rubocop:disable Metrics/BlockLength
+  def template_vars # rubocop:disable Metrics/AbcSize
+    api_keys = [Chef::Datadog.api_key(node)]
+    dd_urls = [node['datadog']['url']]
+    node['datadog']['extra_endpoints'].each do |_, endpoint|
+      next unless endpoint['enabled']
+      api_keys << endpoint['api_key']
+      dd_urls << if endpoint['url']
+                   endpoint['url']
+                 else
+                   node['datadog']['url']
+                 end
+    end
+    {
+      :api_keys => api_keys,
+      :dd_urls => dd_urls
+    }
+  end
   if is_windows
     owner 'Administrators'
     rights :full_control, 'Administrators'
@@ -79,12 +85,14 @@ template agent_config_file do
   else
     owner 'dd-agent'
     group 'root'
-    mode 0640
+    mode '640'
   end
   variables(
-    :api_keys => api_keys,
-    :dd_urls => dd_urls,
-    :extra_config => extra_config
+    if respond_to?(:lazy)
+      lazy { template_vars }
+    else
+      template_vars
+    end
   )
   sensitive true if Chef::Resource.instance_methods(false).include?(:sensitive)
 end