+++ spec/dd-agent_spec.rb
@@ -223,20 +223,49 @@ describe 'datadog::dd-agent' do
   end
 
   context 'allows a string for agent version' do
-    cached(:chef_run) do
-      ChefSpec::SoloRunner.new(
-        :platform => 'ubuntu',
-        :version => '14.10'
-      ) do |node|
-        node.set['datadog'] = {
-          'api_key' => 'somethingnotnil',
-          'agent_version' => '1:5.9.0-1'
-        }
-      end.converge described_recipe
+    context 'on linux' do
+      cached(:chef_run) do
+        ChefSpec::SoloRunner.new(
+          :platform => 'ubuntu',
+          :version => '14.10'
+        ) do |node|
+          node.set['datadog'] = {
+            'api_key' => 'somethingnotnil',
+            'agent_version' => '1:5.9.0-1'
+          }
+        end.converge described_recipe
+      end
+
+      it 'installs agent 1:5.9.0-1' do
+        expect(chef_run).to install_apt_package('datadog-agent').with(version: '1:5.9.0-1')
+      end
     end
 
-    it 'installs agent 1:5.9.0-1' do
-      expect(chef_run).to install_apt_package('datadog-agent').with(version: '1:5.9.0-1')
+    context 'on windows' do
+      cached(:chef_run) do
+        set_env_var('ProgramData', 'C:\ProgramData')
+        ChefSpec::SoloRunner.new(
+          :platform => 'windows',
+          :version => '2012R2',
+          :file_cache_path => 'C:/chef/cache'
+        ) do |node|
+          node.set['datadog'] = {
+            'api_key' => 'somethingnotnil',
+            'agent_version' => '5.10.1'
+          }
+        end.converge described_recipe
+      end
+
+      temp_file = ::File.join('C:/chef/cache', 'ddagent-cli.msi')
+
+      it_behaves_like 'windows Datadog Agent'
+      # remote_file source gets converted to an array, so we need to do
+      # some tricky things to be able to regex against it
+      # Relevant: http://stackoverflow.com/a/12325983
+      it 'installs agent 5.10.1' do
+        expect(chef_run.remote_file(temp_file).source.to_s)
+          .to match(/ddagent-cli-5.10.1.msi/)
+      end
     end
   end
 
@@ -373,7 +402,9 @@ describe 'datadog::dd-agent' do
 
       it_behaves_like 'debianoids repo'
     end
+  end
 
+  context 'datadog.conf configuration' do
     context 'allows a string for tags' do
       cached(:chef_run) do
         ChefSpec::SoloRunner.new(
@@ -496,6 +527,67 @@ describe 'datadog::dd-agent' do
         expect(chef_run).to render_file('/etc/dd-agent/datadog.conf')
           .with_content(/^api_key: something1,something2$/)
           .with_content(%r{^dd_url: https://app.example.com,https://app.example.com$})
+      end
+    end
+
+    context 'with no api_key set' do
+      cached(:chef_run) do
+        ChefSpec::SoloRunner.new(
+          platform: 'ubuntu',
+          version: '12.04'
+        ) do |node|
+          node.set['languages'] = { 'python' => { 'version' => '2.6.2' } }
+        end.converge described_recipe
+      end
+
+      it 'raises an error' do
+        expect(chef_run).to run_ruby_block('datadog-api-key-unset')
+        expect { chef_run.ruby_block('datadog-api-key-unset').old_run_action(:run) }.to raise_error(RuntimeError)
+      end
+    end
+
+    context 'with api_key set as node attribute and on node run_state' do
+      cached(:chef_run) do
+        ChefSpec::SoloRunner.new(
+          platform: 'ubuntu',
+          version: '12.04'
+        ) do |node|
+          node.set['datadog'] = {
+            'api_key' => 'as_node_attribute',
+            'url' => 'https://app.example.com'
+          }
+          node.set['languages'] = { 'python' => { 'version' => '2.6.2' } }
+          node.run_state['datadog'] = { 'api_key' => 'on_run_state' }
+        end.converge described_recipe
+      end
+
+      it_behaves_like 'common linux resources'
+
+      it 'uses the api_key from the run_state' do
+        expect(chef_run).to render_file('/etc/dd-agent/datadog.conf')
+          .with_content(/^api_key: on_run_state$/)
+
+        expect(chef_run).not_to render_file('/etc/dd-agent/datadog.conf')
+          .with_content(/api_key: as_node_attribute/)
+      end
+    end
+
+    context 'with api_key set on node run_state only' do
+      cached(:chef_run) do
+        ChefSpec::SoloRunner.new(
+          platform: 'ubuntu',
+          version: '12.04'
+        ) do |node|
+          node.set['languages'] = { 'python' => { 'version' => '2.6.2' } }
+          node.run_state['datadog'] = { 'api_key' => 'on_run_state' }
+        end.converge described_recipe
+      end
+
+      it_behaves_like 'common linux resources'
+
+      it 'uses the api_key from the run_state' do
+        expect(chef_run).to render_file('/etc/dd-agent/datadog.conf')
+          .with_content(/^api_key: on_run_state$/)
       end
     end
   end